backend/requirements.txt
flask
flask-cors
kafka-python
python-dotenv
opencv-python
requests
argparse
torch
numpy
scikit-learn
mediapipe
importlib-metadata
pandas

# Optional dependencies based on platform
# On Raspberry Pi, install these packages:
# picamera2
# libcamera